imperium-blitz/requirements.txt
mss
pygetwindow
Pillow
